requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pandas
scikit-learn
matplotlib
seaborn
shap
python-multipart